// src/app/features/feature-list-management/models/user-menu.model.ts
export interface UserMenu {
  ModuleId: string;
  ModuleName: string;
  Level: number;
  ParentId: string;
  Url: string;
  Remark: string;
  Status: string;
  OrderSeq: number;
  ButtonList: string;
  ModDate: string;
  ModTime: string;
  ModUser: string;
  A: string;
  D: string;
  L: string;
  I: string;
  U: string;
}